refactor: split utils in various files
