+++ src/components/draw/particles.ts
@@ -3,7 +3,7 @@ import { getKaboomContext } from "../../kaboom";
 import { lerp, map, Quad, Vec2, vec2 } from "../../math";
 import { Color } from "../../math/color";
 import type { Comp, KaboomCtx, ShapeType, Vertex } from "../../types";
-import { KEvent } from "../../utils";
+import { KEvent } from "../../utils/";
 
 class Particle {
     pos: Vec2 = vec2(0);
