+++ src/components/transform/offscreen.ts
@@ -2,7 +2,7 @@ import { DEF_OFFSCREEN_DIS } from "../../constants";
 import { getKaboomContext } from "../../kaboom";
 import { Rect, vec2 } from "../../math";
 import type { Comp, GameObj, KaboomCtx } from "../../types";
-import type { KEventController } from "../../utils";
+import type { KEventController } from "../../utils/";
 import type { PosComp } from "./pos";
 
 /**
