+++ src/kaboom.ts
@@ -121,7 +121,7 @@ import {
     Registry,
     runes,
     uid,
-} from "./utils";
+} from "./utils//";
 
 import { FontData } from "./gfx/fonts";
