+++ src/utils/asserts.ts
@@ -12,3 +12,9 @@ export const arrayIsVec2 = (arr: unknown[]): arr is Vec2[] => {
 export const arrayIsNumber = (arr: unknown[]): arr is number[] => {
     return typeof arr[0] === "number";
 };
+
+export function isClass(obj: any): obj is Function {
+    return obj?.prototype
+        && Object.getOwnPropertyDescriptor(obj.prototype, "constructor")
+            !== undefined;
+}
