+++ src/utils/uid.ts
@@ -1,0 +1,4 @@
+export const uid = (() => {
+    let id = 0;
+    return () => id++;
+})();
